[6889] update migration tasks to entities

+++ hapi-fhir-jpaserver-model/src/main/java/ca/uhn/fhir/jpa/model/entity/IBaseResourceEntity.java
@@ -48,7 +48,7 @@ public interface IBaseResourceEntity<T> {
 
 	String getResourceType();
 
-	Integer getResourceTypeId();
+	Short getResourceTypeId();
 
 	InstantDt getUpdated();
